Restructured and cleaned up linear IR into packages.
